refactor(ecommerce): change directory structure
