+++ src/main/kotlin/ecommerce/controller/ProductViewController.kt
@@ -1,5 +1,7 @@
-package ecommerce
+package ecommerce.controller
 
+import ecommerce.repository.ProductRepository
+import ecommerce.model.Product
 import org.springframework.stereotype.Controller
 import org.springframework.ui.Model
 import org.springframework.web.bind.annotation.GetMapping
